Refactor sample tests configuration to reduce the efforts of adding samples. (#1730)
* Remove redundant import.

* Simplify sample_test.yaml by using withItem syntax.

* Simplify sample_test.yaml by using withItem syntax.

* Change dict to str in withItems.

* remove redundant sed options.

* Fix format/style issues

* [WIP] Refactor repeated logic into two utility functions.

* [WIP] Add a utility function to validate the test results from a notebook test.

* [WIP] Refactor test cases (except for notebook sample tests) into adopting utility functions.
TODO: Need to move the functions of run_*_test.py into a unified run_sample_test.py.

* [WIP] Fix a typo in test name and incorporate tfx-cab-classification, kubeflow-training-classification, xgboost-training-cm and basic ones into one run_sample_test.py

* Fix/add some comments.

* Refactor notebook tests into using utility functions

* lint

* Unify naming in sample_test.yaml

* Remove old *_test.py files

* Fix tests by fixing test names.

* Fix string formatting, per Ark-kun's comment.

* Fix names of papermill-generated python notebook.

* Fix tests

* Fix test by fixing experiment names, and test names in yaml.

* Fix test by fixing experiment names.

* Fix dsl type checking test that does not require experiment set-up.

* Remove redundant commands and usage of ipython

* Revert "Remove redundant commands and usage of ipython"

This reverts commit 23a0e014

* Remove redundant string subs and edit an AI.

* Move image name injection to a utility function to improve readability.

* Revert lint changes of check_notebook_results.py

* Unify test case naming convention to underscore.

* Fix .py name

* Fix README.md

* Fix test

* Add TODO items.

* Add a utility function to inject kubeflow_training_classification python sample file.

* Fix redundant cd command.

* Fix indentation.

* Fix test names in component_test.yaml

* Remove redundant clean_cmle_models.py

* Fix nit problem.

* Fix comment.
