+++ samples/core/kubeflow_training_classification/kubeflow_training_classification.py
