+++ samples/core/tfx_cab_classification/README.md
@@ -2,9 +2,11 @@ The `taxi-cab-classification-pipeline.py` sample runs a pipeline with TensorFlow
 
 ## The dataset
 
-This sample is based on the model-analysis example [here](https://github.com/tensorflow/model-analysis/tree/master/examples/chicago_taxi).
+This sample is based on the model-analysis example 
+[here](https://github.com/tensorflow/tfx/tree/master/tfx/examples/chicago_taxi).
 
-The sample trains and analyzes a model based on the [Taxi Trips dataset](https://data.cityofchicago.org/Transportation/Taxi-Trips/wrvz-psew)
+The sample trains and analyzes a model based on the 
+[Taxi Trips dataset](https://data.cityofchicago.org/Transportation/Taxi-Trips/wrvz-psew)
 released by the City of Chicago.
 
 Note: This site provides applications using data that has been modified
@@ -54,7 +56,10 @@ Open the Kubeflow pipelines UI. Create a new pipeline, and then upload the compi
   1. The name of a GCP project.
   2. An output directory in a Google Cloud Storage bucket, of     the form `gs://<BUCKET>/<PATH>`.
 - On-Premise
-  For On-Premise cluster, the pipeline will create a Persistent Volume Claim (PVC), and download automatically the [source date](https://github.com/kubeflow/pipelines/tree/master/samples/tfx/taxi-cab-classification) to the PVC.
+  For On-Premise cluster, the pipeline will create a Persistent Volume Claim (PVC), and download 
+  automatically the 
+  [source data](https://github.com/kubeflow/pipelines/tree/master/samples/core/tfx_cab_classification/taxi-cab-classification) 
+  to the PVC.
   1. The `output` is PVC mount point for the containers, can be set to `/mnt`.
   2. The `project` can be set to `taxi-cab-classification-pipeline-onprem`.
   3. If the PVC mounted to `/mnt`, the value of below parameters need to be set as following:
