+++ samples/core/tfx_cab_classification/taxi-cab-classification/column-names.json
