+++ samples/core/tfx_cab_classification/taxi-cab-classification/schema.json
