+++ samples/core/tfx_cab_classification/tfx_cab_classification.py
